-rename method 'remove' to 'removeByValue' -rename test method
