--- src/test/java/org/example/PersonForTest.java
@@ -1,35 +1,0 @@
-package org.example;
-
-public class PersonForTest {
-    private String name;
-    private int age;
-
-    public PersonForTest (String name, int age) {
-        this.name = name;
-        this.age = age;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-
-    @Override
-    public String toString() {
-        return "Person{" +
-                "name='" + name + '\'' +
-                ", age=" + age +
-                '}';
-    }
-}
